feat: available rooms by dates and blocked

+++ src/main/java/com/roomify/roomifybackend/persistence/entity/RoomTypeEntity.java
@@ -32,7 +32,7 @@ public class RoomTypeEntity {
     @CollectionTable(name = "room_image", joinColumns = @JoinColumn(name = "room_id"))
     @Column(name = "image_url")
     private List<String> images;
-    private Long quantity_available;
+    private int quantity_available;
     private int capacity;
     private BigDecimal price;
