+++ src/main/java/com/roomify/roomifybackend/presentation/dto/response/RoomTypeResponse.java
@@ -14,7 +14,7 @@ public record RoomTypeResponse(
         int bathRooms,
         int meters,
         List<String> images,
-        Long quantity_available,
+        int quantity_available,
         int capacity,
         BigDecimal price,
         Set<AmenityEntity> amenities) {
